add account utility summary table
