+++ src/app/overview-report/account-report/account-report-utility-usage-table/account-report-utility-usage-table.component.css
@@ -1,0 +1,7 @@
+.w-1{
+    width: 10%;
+}
+
+.w-2{
+    width: 15%;
+}
