+++ src/app/overview-report/account-report/account-report-utility-usage-table/account-report-utility-usage-table.component.ts
@@ -1,20 +1,21 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
-import { IdbFacility, IdbUtilityMeter } from 'src/app/models/idb';
-import { OverviewReportService, ReportOptions, ReportUtilityOptions, ReportUtilitySummary } from '../../overview-report.service';
+import { IdbAccount, IdbUtilityMeter } from 'src/app/models/idb';
+import { OverviewReportService, ReportUtilityOptions, ReportUtilitySummary } from '../../overview-report.service';
 
 @Component({
-  selector: 'app-facility-report-utility-usage-table',
-  templateUrl: './facility-report-utility-usage-table.component.html',
-  styleUrls: ['./facility-report-utility-usage-table.component.css']
+  selector: 'app-account-report-utility-usage-table',
+  templateUrl: './account-report-utility-usage-table.component.html',
+  styleUrls: ['./account-report-utility-usage-table.component.css']
 })
-export class FacilityReportUtilityUsageTableComponent implements OnInit {
+export class AccountReportUtilityUsageTableComponent implements OnInit {
   @Input()
-  facility: IdbFacility;
+  account: IdbAccount;
+  
   reportUtilityOptions: ReportUtilityOptions;
   reportUtilityOptionsSub: Subscription;
-  facilityReportUtilitySummary: ReportUtilitySummary;
+  accountReportUtilitySummary: ReportUtilitySummary;
   constructor(private utilityMeterDbService: UtilityMeterdbService, private overviewReportService: OverviewReportService) { }
 
   ngOnInit(): void {
@@ -30,7 +31,6 @@ export class FacilityReportUtilityUsageTableComponent implements OnInit {
 
   setFacilitySummary() {
     let accountMeters: Array<IdbUtilityMeter> = this.utilityMeterDbService.accountMeters.getValue();
-    let facilityMeters: Array<IdbUtilityMeter> = accountMeters.filter(meter => { return meter.facilityId == this.facility.id });
-    this.facilityReportUtilitySummary = this.overviewReportService.getUtilityUsageData(facilityMeters, this.reportUtilityOptions, false);
+    this.accountReportUtilitySummary = this.overviewReportService.getUtilityUsageData(accountMeters, this.reportUtilityOptions, true);
   }
 }
